requirements.txt
pandas==1.4.1
requests==2.27.1
pyyaml==6.0
numpy==1.22.3